requirements.txt
Django==4.2.7
Pillow==10.0.1
python-decouple==3.8
whitenoise==6.5.0
gunicorn==21.2.0
psycopg2-binary==2.9.7
django-crispy-forms==2.0
crispy-bootstrap5==0.7
django-ckeditor==6.7.0